prevent a user from updating or deleting another user's event

+++ routes/events.py
@@ -30,6 +30,7 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 @event_router.post("/new")
 async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {"message": "Event created successfully"}
 
@@ -38,6 +39,11 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
 async def update_event(
     id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)
 ) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Operation not allowed"
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
@@ -49,6 +55,11 @@ async def update_event(
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST, detail="Operation not allowed"
+        )
     event = await event_database.delete(id)
     if not event:
         raise HTTPException(
